backend/requirements.txt
flask==3.0.0
flask-cors
google-generativeai
google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-firestore
google-cloud-storage
gunicorn
dateparser
tzdata
